requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
# requirements.txt
apache-airflow-providers-http
apache-airflow-providers-postgres
pandas
